custom exception (#42)
* mapping

* refactor

* loan wrapper

* staged

* changes

* testing

* controller testing

* changes in testing

* changes in testing

* exception

* exception

* exception

* exception

* exception

+++ backend/src/main/java/com/example/backend/service/EmployeeService.java
@@ -1,11 +1,14 @@
 package com.example.backend.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.example.backend.exception.ResourceNotFoundException;
 import com.example.backend.model.Employee;
+import com.example.backend.model.Loan;
 import com.example.backend.repository.EmployeeCardRepository;
 import com.example.backend.repository.EmployeeIssueRepository;
 import com.example.backend.repository.EmployeeRepository;
@@ -29,16 +32,25 @@ public class EmployeeService {
 		return employeeRepository.save(e);
 	}
 	
-	public String deleteEmployee(String id) {
-		System.out.println(employeeRepository.findById(id));
-		if(!employeeRepository.findById(id).isPresent()) return "Failure";
-		employeeIssueRepository.deleteAllByEmployeeId(id);
-		employeeCardRepository.deleteAllByEmployeeId(id);
-		employeeRepository.deleteById(id);
-		return "Success";
+	public String deleteEmployee(String id) throws ResourceNotFoundException{
+		Optional<Employee> em = employeeRepository.findById(id);
+		if(em!=null) {
+			employeeIssueRepository.deleteAllByEmployeeId(id);
+			employeeCardRepository.deleteAllByEmployeeId(id);
+			employeeRepository.deleteById(id);
+			return "Success";
+		}
+		else 
+			throw new ResourceNotFoundException("Employee Id does not exist");
+		
 	}
-	public Employee updateEmployee(Employee e) {
-		employeeRepository.deleteById(e.getEmployee_id());
-		return employeeRepository.save(e);
+	public Employee updateEmployee(Employee e) throws ResourceNotFoundException {
+		Optional<Employee> em = employeeRepository.findById(e.getEmployee_id());
+		if(em!=null) {
+			employeeRepository.deleteById(e.getEmployee_id());
+			return employeeRepository.save(e);
+		}
+		else 
+			throw new ResourceNotFoundException("Employee Id does not exist");
 	}
 }
