+++ backend/src/main/java/com/example/backend/service/ItemService.java
@@ -4,6 +4,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.example.backend.exception.NoDataFoundException;
+
+import com.example.backend.exception.ResourceNotFoundException;
+import com.example.backend.model.Employee;
+
 import com.example.backend.model.EmployeeCard;
 import com.example.backend.model.EmployeeIssue;
 import com.example.backend.model.Item;
@@ -36,31 +40,45 @@ public class ItemService {
 	@Autowired
 	LoanRepository loanRepository;
 	
-	public List<Item> getAllItems(){
-		return itemRepository.findAll();
+	public List<Item> getAllItems() throws NoDataFoundException{
+		List<Item> list = new ArrayList<>();
+		
+		list= itemRepository.findAll();
+		if(list.size()!=0)
+			return list;
+		else
+			throw new NoDataFoundException("No items found");
 	}
 	
 	@Transactional
-	public List<Object> apply(Item item, String id) {
+	public List<Object> apply(Item item, String id) throws ResourceNotFoundException {
 		List<Object> list = new ArrayList<>();
 		String type = item.getItemCategory();
 		Loan loan = loanRepository.findByLoanType(type);
 		int duration = loan.getDuration();
-		
+		Optional<Employee> emp = employeeRepository.findById(id);
+    	if(emp!=null)
+    	{
+    		EmployeeIssue employeeIssue = new EmployeeIssue(new Date(),this.getReturnDate(new Date(), duration),id,item.getItemId());
+    		list.add(employeeIssueRepository.save(employeeIssue));
+    		EmployeeCard employeeCard = new EmployeeCard(id,loan.getLoan_id(),new Date());
+    		list.add(employeeCardRepository.save(employeeCard));
+    	
+    		return list;
+    	}
+    	else 
+    		throw new ResourceNotFoundException("Employee ID does not exist");
 
-		EmployeeIssue employeeIssue = new EmployeeIssue(new Date(),this.getReturnDate(new Date(), duration),id,item.getItemId());
-		list.add(employeeIssueRepository.save(employeeIssue));
-		EmployeeCard employeeCard = new EmployeeCard(id,loan.getLoan_id(),new Date());
-		list.add(employeeCardRepository.save(employeeCard));
-	
-		return list;		
+				
 	}
 	
-	public List<PurchasedItem> getItemsById(String id) throws NoDataFoundException{
+
+	public List<PurchasedItem> getItemsById(String id) throws ResourceNotFoundException{
 		List<PurchasedItem> list= new ArrayList<>();
 		employeeIssueRepository.findByEmployeeId(id).forEach((obj)-> list.add(new PurchasedItem(obj.getIssue_id(),itemRepository.findById(obj.getItemId()).get())));
 		if(list.size()==0)
-			throw new NoDataFoundException("Employee Id not found ");
+			throw new ResourceNotFoundException("Employee Id not found ");
+
 		else 
 			return list;
 		
@@ -72,21 +90,41 @@ public class ItemService {
 		return calendar.getTime();
 	}
 	
-	public List<Item> findAllItems(){
-		return itemRepository.findAll();
+	public List<Item> findAllItems() throws NoDataFoundException{
+		List<Item> list = new ArrayList<Item>();
+		list=itemRepository.findAll();
+		if(list.size()!=0)
+			return list;
+		else 
+			throw new NoDataFoundException("No items found");
 	}
 	
 	public Item addItem(Item item) {
 		return itemRepository.save(item);
 	}
 	
-	public String deleteItem(String id) {
-		employeeIssueRepository.deleteAllByItemId(id);
-		itemRepository.deleteById(id);
-		return "Success";
+	public String deleteItem(String id) throws ResourceNotFoundException {
+		Optional<Loan> loan = loanRepository.findById(id);
+    	if(loan!=null)
+		{
+    		employeeIssueRepository.deleteAllByItemId(id);
+    		itemRepository.deleteById(id);
+    		return "Success";
+		}
+    	else 
+    		throw new ResourceNotFoundException("Itemn Id does not exist");
+		
 	}
-	public Item updateItem(Item item) {
-		itemRepository.deleteById(item.getItemId());
-		return itemRepository.save(item);
+	public Item updateItem(Item item) throws ResourceNotFoundException {
+		Optional<Item> it = itemRepository.findById(item.getItemId());
+		if(it!=null)
+		{
+			itemRepository.deleteById(item.getItemId());
+			return itemRepository.save(item);
+		}
+		else
+			throw new ResourceNotFoundException("Item id does not exist");
+		
+		
 	}
 }
